+++ wezterm-gui/src/termwindow/mod.rs
@@ -1161,6 +1161,12 @@ impl TermWindow {
                 MuxNotification::SaveToDownloads { .. } => {
                     // Handled by frontend
                 }
+                MuxNotification::PaneFocused(_) => {
+                    // Handled by clientpane
+                }
+                MuxNotification::TabResized(_) => {
+                    // Handled by wezterm-client
+                }
                 MuxNotification::PaneAdded(_)
                 | MuxNotification::PaneRemoved(_)
                 | MuxNotification::WindowWorkspaceChanged(_)
@@ -1345,6 +1351,8 @@ impl TermWindow {
             }
             | MuxNotification::AssignClipboard { .. }
             | MuxNotification::SaveToDownloads { .. }
+            | MuxNotification::PaneFocused(_)
+            | MuxNotification::TabResized(_)
             | MuxNotification::PaneRemoved(_)
             | MuxNotification::WindowCreated(_)
             | MuxNotification::ActiveWorkspaceChanged(_)
@@ -1980,6 +1988,8 @@ impl TermWindow {
             .get_window_mut(self.mux_window_id)
             .ok_or_else(|| anyhow!("no such window"))?;
 
+        // This logic is coupled with the CliSubCommand::ActivateTab
+        // logic in wezterm/src/main.rs. If you update this, update that!
         let max = window.len();
 
         let tab_idx = if tab_idx < 0 {
@@ -2012,6 +2022,8 @@ impl TermWindow {
         let max = window.len();
         ensure!(max > 0, "no more tabs");
 
+        // This logic is coupled with the CliSubCommand::ActivateTab
+        // logic in wezterm/src/main.rs. If you update this, update that!
         let active = window.get_active_idx() as isize;
         let tab = active + delta;
         let tab = if wrap {