+++ wezterm-mux-server-impl/src/dispatch.rs
@@ -159,6 +159,18 @@ where
                     stream.flush().await.context("flushing PDU to client")?;
                 }
             }
+            Ok(Item::Notif(MuxNotification::PaneFocused(pane_id))) => {
+                Pdu::PaneFocused(codec::PaneFocused { pane_id })
+                    .encode_async(&mut stream, 0)
+                    .await?;
+                stream.flush().await.context("flushing PDU to client")?;
+            }
+            Ok(Item::Notif(MuxNotification::TabResized(tab_id))) => {
+                Pdu::TabResized(codec::TabResized { tab_id })
+                    .encode_async(&mut stream, 0)
+                    .await?;
+                stream.flush().await.context("flushing PDU to client")?;
+            }
             Ok(Item::Notif(MuxNotification::ActiveWorkspaceChanged(_))) => {}
             Ok(Item::Notif(MuxNotification::Empty)) => {}
             Err(err) => {